pamangan.com/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
google-generativeai==0.8.3
groq==0.13.1
gunicorn==22.0.0
PyJWT==2.9.0
